requirements.txt
fastapi
gunicorn
uvicorn
websockets
pymongo
python-dotenv
argon2-cffi